pyproject.toml
[tool.poetry]
name = "fastapi-course"
version = "0.1.0"
description = ""
authors = ["Andreas Atle <atle.andreas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"
faker = "^24.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^8.1.1"
httpx = "^0.27.0"
bcrypt = "^4.1.2"
aiofiles = "^23.2.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
